Add eventDataChecksum to the Duty Status Object
OTAPI shouldn't have to  do the calculation, but we should be able to receive the output of the ELD's validation process - i.e. a valid checksum. I think it would be a nice-to-have for OTAPI to verify the result of the ELD event log against Table 3 of the ELD mandate.